Package name refactoring
* Moved content from  Client/Server/Command packages, to RPC
