+++ src/RPCLibrary/RPCLibrary.csproj
@@ -12,4 +12,8 @@
     <PackageReference Include="NLua" Version="1.7.3" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="DataProtocol\" />
+  </ItemGroup>
+
 </Project>
